task1/resolve/requirements.txt
pillow==11.3.0
matplotlib==3.10.6
numpy==2.2.6
pydantic==2.10.6
pydantic-settings==2.7.1
FreeSimpleGUI==5.2.0.post1
opencv-python==4.12.0.88